+++ Libraries/MBZ.AdventOfCode.Core/Solvers/PostPuzzleSolverRunResult.cs
@@ -1,6 +1,6 @@
 ﻿namespace MBZ.AdventOfCode.Core.Solvers;
 
-public record PostPuzzleSolverRunResult(int Expected, int Result)
+public record PostPuzzleSolverRunResult(long Expected, long Result)
 {
     public bool Success => Expected == Result;
 }
